requirements.txt
torch>=2.1.0
transformers>=4.40.0
peft>=0.10.0
bitsandbytes>=0.43.0
datasets>=2.18.0
accelerate>=0.29.0
trl>=0.8.0
scipy>=1.12.0
numpy>=1.26.0
pyyaml>=6.0
pygame>=2.5.0
tqdm>=4.66.0
scikit-learn>=1.4.0
rouge-score>=0.1.2
